requirements.txt
pandas
numpy
requests
sqlalchemy
pymysql
google-cloud-bigquery
streamlit
python-dotenv